# docs/api/backend-openapi.yaml
openapi: 3.1.0
info:
  title: LQ.AI Backend API
  version: 0.1.0
  description: |
    Canonical API specification for the LQ.AI backend (FastAPI service).
    
    This is a **sketch** — it covers the endpoints, request/response shapes, and
    schemas needed to anchor implementation. Expect refinement during M1 build.
    
    The Inference Gateway has a separate OpenAPI specification at
    `gateway-openapi.yaml` (consumes OpenAI-compatible inference requests from
    this backend; not directly accessible to end-users).
    
    Conventions:
    - All endpoints under `/api/` are user-facing; under `/admin/` require admin role.
    - Authentication via Bearer JWT in `Authorization` header unless noted.
    - All resource IDs are UUIDs.
    - All timestamps are ISO 8601 with timezone (UTC by default).
    - Pagination uses cursor-based pagination via `cursor` and `limit` query params.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: LQ.AI
    url: https://github.com/legalquants/lq-ai
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://{operator-host}/api
    description: Operator-deployed instance
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Authenticate with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
        '423':
          description: MFA required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaChallenge'
  /api/v1/auth/mfa/verify:
    post:
      tags: [auth]
      summary: Complete MFA challenge with TOTP code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfa_token, code]
              properties:
                mfa_token: {type: string, description: "Token from the MFA challenge response"}
                code: {type: string, pattern: '^\d{6}$'}
      responses:
        '200':
          description: MFA verification successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid MFA code
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string}
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
...
  /api/v1/auth/change-password:
    post:
      tags: [auth]
      summary: Change the calling user's password
      description: |
        Bearer-authenticated. Verifies the current password before
        accepting a new one (defense against a stolen access token).
        Clears `must_change_password` if it was set (used by the
        first-run admin flow per Task B2). Revokes all active sessions
        for the user; the caller must re-authenticate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204': {description: Password changed successfully}
        '400': {description: New password fails policy (too short, identical to current)}
        '401': {description: Bearer token invalid or current_password is wrong}
        '403':
          description: |
            Reserved by the must-change-password gate elsewhere; the
            change-password endpoint itself is reachable while the gate
            is active.

  # ----------- USERS -----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token: {type: string}
        token_type: {type: string, enum: [Bearer]}
        expires_in: {type: integer, description: "Lifetime in seconds"}
        refresh_token: {type: string}
        user: {$ref: '#/components/schemas/User'}
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: {type: string, description: "Short-lived JWT access token (~15 min)."}
        refresh_token:
          type: string
          description: |
            Refresh token returned on /auth/login and rotated on every
            /auth/refresh per PRD §5.1. The bearer of the most recent
            refresh_token can mint new access tokens; older refresh
            tokens are revoked when rotation happens.
        token_type: {type: string, enum: [Bearer]}
        expires_in: {type: integer, description: "Seconds until the access token expires."}
    MfaChallenge:
      type: object
      required: [mfa_token, methods]
      properties:
        mfa_token: {type: string}
        methods: {type: array, items: {type: string, enum: [totp, recovery_code]}}
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          format: password
          description: |
            The user's current password. Required even when the
            `must_change_password` flag is set — a stolen access token
            alone must not be sufficient to rotate credentials.
        new_password:
          type: string
          format: password
          description: |
            The replacement password. Must satisfy the deployment's
            password policy (minimum length, etc.) and must differ from
            `current_password`. The first-run admin flow accepts any
            password meeting policy; subsequent rotations may add
            complexity rules.
    User:
      type: object
      required: [id, email, is_admin, mfa_enabled, must_change_password, created_at]
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        display_name: {type: string, nullable: true}
        is_admin: {type: boolean}
        mfa_enabled: {type: boolean}
        must_change_password:
          type: boolean
          description: |
            When true, the user must call POST /api/v1/auth/change-password
            before they can use any other authenticated endpoint. Set on
            the auto-created first-run admin (Task B2) and on accounts
            after `reset-admin-password`. Other authenticated endpoints
            return 403 with `error.code = "password_change_required"`
            until this clears.
        created_at: {type: string, format: date-time}
        last_login_at: {type: string, format: date-time, nullable: true}
...
security:
  - bearerAuth: []